% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUrl.R
\name{readUrl}
\alias{readUrl}
\title{Function with try-cath to read a raw csv table from url}
\usage{
readUrl(url)
}
\arguments{
\item{url}{desired url to raw format data}
}
\value{
csv table as a dataframe
}
\description{
Function with try-cath to read a raw csv table from url
}
